+++ src/components/Tabs.tsx
@@ -2,6 +2,10 @@ import * as React from 'react';
 import Tabs from '@mui/material/Tabs';
 import Tab from '@mui/material/Tab';
 import Box from '@mui/material/Box';
+import Divider from '@mui/material/Divider';
+import useMediaQuery from '@mui/material/useMediaQuery';
+import { useTheme } from '@mui/material/styles';
+import AboutMe from './tabs/AboutMe';
 
 interface TabPanelProps {
 	children?: React.ReactNode;
@@ -11,6 +15,8 @@ interface TabPanelProps {
 
 function TabPanel(props: TabPanelProps) {
 	const { children, value, index, ...other } = props;
+	const theme = useTheme();
+	const isMobile = useMediaQuery(theme.breakpoints.down('sm'));
 
 	return (
 		<div
@@ -21,7 +27,10 @@ function TabPanel(props: TabPanelProps) {
 			{...other}
 		>
 			{value === index && (
-				<Box sx={{ p: 3, height: '100%' }}>
+				<Box sx={{
+					pl: isMobile ? 0 : 3,
+					pt: isMobile ? 3 : 0
+				}}>
 					{children}
 				</Box>
 			)}
@@ -36,95 +45,116 @@ function a11yProps(index: number) {
 	};
 }
 
-export default function VerticalTabs() {
+export default function BasicTabs() {
 	const [value, setValue] = React.useState(0);
+	const theme = useTheme();
+	const isMobile = useMediaQuery(theme.breakpoints.down('sm')); // Check if screen size is mobile
 
 	const handleChange = (event: React.SyntheticEvent, newValue: number) => {
 		setValue(newValue);
 	};
 
 	return (
-		<Box sx={{ flexGrow: 1, display: 'flex', height: '100%' }}>
-			<Box sx={{ display: 'flex', flexDirection: 'column', flexShrink: 0 }}>
-				<Tabs
-					orientation="vertical"
-					value={value}
-					onChange={handleChange}
-					aria-label="Vertical tabs example"
-					sx={{ borderRight: 1, borderColor: 'divider', minWidth: '125px' }}
-				>
-					<Tab
-						label="About Me"
-						sx={{
-							color: 'white',
+		<Box sx={{ flexGrow: 1, display: 'flex', flexDirection: isMobile ? 'column' : 'row' }}>
+			<Tabs
+				orientation={isMobile ? 'horizontal' : 'vertical'} // Switch orientation based on screen size
+				value={value}
+				onChange={handleChange}
+				aria-label="Tabs example"
+				sx={{
+					borderBottom: isMobile ? 1 : 0,
+					borderRight: isMobile ? 0 : 1,
+					borderColor: 'divider',
+					minWidth: isMobile ? 'auto' : '125px',
+					display: 'flex',
+					flexDirection: isMobile ? 'row' : 'column', // Adjust direction for mobile
+				}}
+			>
+				<Tab
+					label="About Me"
+					sx={{
+						color: 'white',
+						textAlign: 'left',
+						justifyContent: 'flex-start',
+						flexDirection: 'row',
+						'& .MuiTab-wrapper': {
 							textAlign: 'left',
-							justifyContent: 'flex-start',
-							flexDirection: 'row',
-							'& .MuiTab-wrapper': {
-								textAlign: 'left',
-							},
-						}}
-						{...a11yProps(0)}
-					/>
-					<Tab
-						label="Projects"
-						sx={{
-							color: 'white',
+						},
+					}}
+					{...a11yProps(0)}
+				/>
+				<Tab
+					label="Projects"
+					sx={{
+						color: 'white',
+						textAlign: 'left',
+						justifyContent: 'flex-start',
+						flexDirection: 'row',
+						'& .MuiTab-wrapper': {
 							textAlign: 'left',
-							justifyContent: 'flex-start',
-							flexDirection: 'row',
-							'& .MuiTab-wrapper': {
-								textAlign: 'left',
-							},
-						}}
-						{...a11yProps(1)}
-					/>
-					<Tab
-						label="Experience"
-						sx={{
-							color: 'white',
+						},
+					}}
+					{...a11yProps(1)}
+				/>
+				<Tab
+					label="Experience"
+					sx={{
+						color: 'white',
+						textAlign: 'left',
+						justifyContent: 'flex-start',
+						flexDirection: 'row',
+						'& .MuiTab-wrapper': {
 							textAlign: 'left',
-							justifyContent: 'flex-start',
-							flexDirection: 'row',
-							'& .MuiTab-wrapper': {
-								textAlign: 'left',
-							},
-						}}
-						{...a11yProps(2)}
-					/>
-					<Tab
-						label="Contact"
-						sx={{
-							color: 'white',
+						},
+					}}
+					{...a11yProps(2)}
+				/>
+				<Tab
+					label="Contact"
+					sx={{
+						color: 'white',
+						textAlign: 'left',
+						justifyContent: 'flex-start',
+						flexDirection: 'row',
+						'& .MuiTab-wrapper': {
 							textAlign: 'left',
-							justifyContent: 'flex-start',
-							flexDirection: 'row',
-							'& .MuiTab-wrapper': {
-								textAlign: 'left',
-							},
-						}}
-						{...a11yProps(3)}
-					/>
-				</Tabs>
-			</Box>
+						},
+					}}
+					{...a11yProps(3)}
+				/>
+			</Tabs>
 			<Box sx={{ flexGrow: 1, display: 'flex', flexDirection: 'column' }}>
 				<TabPanel value={value} index={0}>
-					<h2 className="text-white m-0">About Me</h2>
-					<hr />
-					<div className="text-white">
-						Hello! I&apos;m a recent graduate 🎓 of Pace University with a degree in Computer Science. I currently work as a web-developer for Phormulary and part time as a STEP Instructor at Iona University.
-					</div>
+					<AboutMe />
 				</TabPanel>
 				<TabPanel value={value} index={1}>
-					Item Two
+					<h2 className="text-white m-0">Projects</h2>
+					<Divider sx={{ bgcolor: 'white' }} className='my-2' />
+					<div className='text-white'>
+						Under Construction. Projects go here.
+					</div>
 				</TabPanel>
 				<TabPanel value={value} index={2}>
-					Item Three
+					<h2 className="text-white m-0">Experience</h2>
+					<Divider sx={{ bgcolor: 'white' }} className='my-2' />
+					<div className='text-white'>
+						Under Construction. Experiences go here.
+					</div>
 				</TabPanel>
 				<TabPanel value={value} index={3}>
-					Item Four
+					<h2 className="text-white m-0">Contact</h2>
+					<Divider sx={{ bgcolor: 'white' }} className='my-2' />
+					<div className='text-white'>In the meantime, in case you need it:
+						<ul>
+							<li><a href="mailto:okurakeng@gmail.com">Email</a> (📧)</li>
+							<li><a href="https://www.linkedin.com/in/kenji-okura/">Linkedin</a> (👤)</li>
+							<li><a href="https://github.com/OkayKenji/">GitHub Fun</a> (🎉)</li>
+							<li><a href="https://github.com/okurakeng/">GitHub Work</a> (👨‍💻)</li>
+						</ul>
+					</div>
 				</TabPanel>
 			</Box>
 		</Box>
 	);
 }
+
